fix sqrt.rb and triangle.rb

+++ lesson01/triangle_area.rb
@@ -1,7 +1,7 @@
 puts "Input the value of side:"
-a = gets.to_i
+a = gets.to_f
 puts "Input the height:"
-h = gets.to_i
+h = gets.to_f
 
 triangle_area = a / 2 * h 
 puts "The area of your triangle is equal to #{triangle_area}"
